requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
psycopg2-binary>=2.9.9
pillow>=10.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
redis>=5.0.0
httpx>=0.25.0
